requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
SQLAlchemy==2.0.36
alembic==1.13.1
pydantic==2.9.2
python-dotenv==1.0.1
cryptography==43.0.3
httpx==0.27.2
streamlit==1.38.0
rich==13.7.1
click==8.1.7
typer==0.12.3

# Development/testing dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
